+++ packages/astro/src/internal/admin-bar/generated-sources.ts
@@ -246,7 +246,6 @@ const ensureApp = async () => {
 \t\t\tconst wrappedDefinition = configureLucid({
 \t\t\t\truntime: runtimeAdapter,
 \t\t\t\tdb,
-\t\t\t\tenv: envSchema,
 \t\t\t\tconfig: configFactory,
 \t\t\t}, {
 \t\t\t\temailTemplates,
@@ -386,7 +385,6 @@ const ensureApp = async () => {
 \t\t\tconst wrappedDefinition = configureLucid({
 \t\t\t\truntime: runtimeAdapter,
 \t\t\t\tdb,
-\t\t\t\tenv: envSchema,
 \t\t\t\tconfig: configFactory,
 \t\t\t}, {
 \t\t\t\temailTemplates,