+++ packages/astro/src/internal/generated-sources.ts
@@ -78,7 +78,6 @@ const ensureApp = async () => {
 \t\t\tconst wrappedDefinition = configureLucid({
 \t\t\t\truntime: runtimeAdapter,
 \t\t\t\tdb,
-\t\t\t\tenv: envSchema,
 \t\t\t\tconfig: configFactory,
 \t\t\t}, {
 \t\t\t\temailTemplates,
@@ -198,7 +197,6 @@ const ensureToolkit = async () => {
 \t\t\tconst wrappedDefinition = configureLucid({
 \t\t\t\truntime: runtimeAdapter,
 \t\t\t\tdb,
-\t\t\t\tenv: envSchema,
 \t\t\t\tconfig: configFactory,
 \t\t\t}, {
 \t\t\t\temailTemplates,
@@ -331,7 +329,6 @@ const ensureApp = async () => {
 \t\t\tconst wrappedDefinition = configureLucid({
 \t\t\t\truntime: runtimeAdapter,
 \t\t\t\tdb,
-\t\t\t\tenv: envSchema,
 \t\t\t\tconfig: configFactory,
 \t\t\t}, {
 \t\t\t\temailTemplates,
@@ -500,7 +497,6 @@ const ensureToolkit = async () => {
 \t\t\tconst wrappedDefinition = configureLucid({
 \t\t\t\truntime: runtimeAdapter,
 \t\t\t\tdb,
-\t\t\t\tenv: envSchema,
 \t\t\t\tconfig: configFactory,
 \t\t\t}, {
 \t\t\t\temailTemplates,