+++ packages/astro/src/internal/worker-module.ts
@@ -197,7 +197,6 @@ return astroWorker.fetch(request, env, ctx);`,
 	const wrappedDefinition = astroConfigureLucid({
 		runtime,
 		db,
-		env: envSchema,
 		config: configFactory,
 	}, {
 		emailTemplates,