+++ packages/runtime-node/src/cli/build.ts
@@ -106,7 +106,6 @@ const startServer = async () => {
 		const definition = {
 			runtime: runtimeAdapter,
 			db,
-			env: envSchema,
 			config: configFactory,
 		};
 		const wrappedDefinition = runtimeAdapter.configureLucid